# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Rocketium Assignment API",
      "description": "API documentation for Rocketium Assignment project"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "tags": [
      {
        "name": "Data",
        "description": "API for data in the system"
      }
    ],
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
      "/data": {
        "get": {
          "tags": ["Data"],
          "summary": "Get complete data",
          "description": "Retrieve all the data from the server",
          "produces": ["application/json"],
          "responses": {
            "200": {
              "description": "Successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Data"
                }
              }
            }
          }
        }
      },
      "/data/{id}": {
        "get": {
          "tags": ["Data"],
          "summary": "Get data by specific id",
          "description": "Retrieve a specific data item by its ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "string",
              "description": "ID of the data item to retrieve"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful operation",
              "schema": {
                "$ref": "#/definitions/Data"
              }
            },
            "404": {
              "description": "Data not found"
            }
          }
        }
      }
    },
    "definitions": {
      "Data": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "language": {
            "type": "string"
          }
        }
      }
    }
  }
  